requirements.txt
pre-commit==4.0.1
PyYAML==6.0.2
pytest==8.3.5
pytest-cov==6.1.0
coverage==7.8.0
build==1.2.2.post1
